Added comments and GeoComposeState functions
