+++ toolkit/geo-compose/src/main/java/com/arcgismaps/toolkit/geocompose/Map.kt
@@ -73,7 +73,7 @@ public fun Map(modifier: Modifier = Modifier, mapState: MapState = MapState()) {
                     currentViewpoint = mapView.getCurrentViewpoint(ViewpointType.CenterAndScale)
                 }
                 if (currentViewpoint != null) {
-                    mapState.setViewpoint(currentViewpoint)
+                    mapState.setCurrentViewpoint(currentViewpoint)
                 }
             }
         }
